settings supposed to work

+++ src/main/java/ru/mipt/messenger/repositories/SettingsRepository.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 
 import ru.mipt.messenger.models.Settings;
 
-public interface SettingsRepository extends JpaRepository<Settings, Long> {
-    Optional<Settings> findByUserId(Long userId);
+public interface SettingsRepository extends JpaRepository<Settings, Integer> {
+    Optional<Settings> findByUserId(Integer userId);
 }
